added the following fixes:
if the PID we're looking for has gone, we must fail
or we'll go into an infinite loop.

if the xterm fails to launch (eg. eg have no display),
get_tty will return "fail" which we intercept here.

if the xterm launch failed, continue with the next PID in the list.